+++ addon-test-support/fillIn.js
@@ -1,5 +1,5 @@
 import { fillIn as rawFillin } from '@ember/test-helpers';
-import findControl from './find-control';
+import findControl from './finders/control';
 
 export default async function fillIn(label, value) {
   let control = findControl(label);
