+++ addon-test-support/finders/button.js
@@ -1,5 +1,5 @@
-import sharedFind from './utils/shared-find';
-import findButtons from './find-buttons';
+import sharedFind from './shared';
+import findButtons from './buttons';
 
 export default function findButton(labelText) {
   return sharedFind(findButtons, labelText, 'button');
