+++ addon-test-support/finders/buttons.js
@@ -1,5 +1,5 @@
-import findObjects from './find-objects';
-import { buttonQuery } from './dom/selectors'
+import findObjects from './objects';
+import { buttonQuery } from './../dom/selectors'
 
 export default function findButtons(labelText) {
   return findObjects(buttonQuery, labelText, 'button');
