+++ addon-test-support/finders/controls.js
@@ -1,5 +1,5 @@
-import findObjects from './find-objects';
-import { formControlQuery } from './dom/selectors';
+import findObjects from './objects';
+import { formControlQuery } from './../dom/selectors'
 
 export default function findControls(labelText) {
   return findObjects(formControlQuery, labelText, 'control');
