+++ addon-test-support/finders/object.js
@@ -1,0 +1,5 @@
+import findObjects from './find-objects';
+
+export default function (selector, labelText) {
+  let objects = findObjects(selector, labelText)
+}
