+++ addon-test-support/finders/objects.js
@@ -1,4 +1,4 @@
-import findAll from './utils/find-all';
+import findAll from './find-all';
 
 export default function findObjects(selector, labelText) {
   return findAll(selector, labelText)
