+++ addon-test-support/index.js
@@ -2,8 +2,8 @@ import fillIn from './fillIn';
 import select from './select';
 import toggle from './toggle';
 import click from './click';
-import findControl from './find-control';
-import findButton from './find-button';
+import findControl from './finders/control';
+import findButton from './finders/button';
 
 export {
   fillIn,
