+++ addon-test-support/toggle.js
@@ -1,5 +1,5 @@
 import { click } from '@ember/test-helpers';
-import findControl from './find-control';
+import findControl from './finders/control';
 
 export default async function toggle(label) {
   let control = await findControl(label);
